requirements.txt
psycopg2-binary
py4j
pyspark
boto3
aiofiles
PyYAML
